% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc_function.R
\name{aac}
\alias{aac}
\title{Param: Area below the curve}
\usage{
aac(dados, trat, nrep, time)
}
\arguments{
\item{dados}{data.frame containing the responses of the evaluations in separate columns side by side and without the columns with the identification of the factors}

\item{trat}{vector of treatments with n repetitions}

\item{nrep}{Number of repetitions}

\item{time}{vector containing time}
}
\value{
Returns a vector with the index
}
\description{
Calculates the area under the germination or emergence curve. A parameter that can replace the traditional emergence or germination speed index.
}
\examples{
data("substrate")
aac(substrate[,c(3:18)],
      trat = substrate$Trat,
      nrep = 4,
      time = 1:16)
}
