% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aristolochia_data.R
\docType{data}
\name{aristolochia}
\alias{aristolochia}
\title{dataset: aristolochia}
\format{
data.frame containing data set
  \describe{
  \item{\code{trat}}{numeric vector with factor 1}
  \item{\code{germ}}{Numeric vector with germination percentage}
  \item{\code{vel}}{numerical vector with germination speed}
  }
}
\usage{
data("aristolochia")
}
\description{
The data come from an experiment conducted at the Seed Analysis
Laboratory of the Agricultural Sciences Center of the State
University of Londrina, in which five temperatures (15, 20, 25,
30 and 35C) were evaluated in the germination of \emph{Aristolochia elegans}.
The experiment was conducted in a completely randomized
design with four replications of 25 seeds each.
}
\examples{
data(aristolochia)
}
\author{
Hugo Roldi Guariz
}
\keyword{datasets}
