% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correl_function.R
\name{correl}
\alias{correl}
\title{Comparison: correlation between parameters}
\usage{
correl(seeds)
}
\arguments{
\item{seeds}{Object returned in the seeds function}
}
\value{
Returns correlation graphs between parameters calculated by traditional methods and by logistic regression
}
\description{
Correlation between the logistical model and the traditional model
}
\examples{
data("substrate")
a=seeds(substrate[,c(3:18)],
      trat = substrate$Trat,
      nrep = 4,
      time = 1:16)
correl(a)
}
