% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substrate_dataset.R
\docType{data}
\name{substrate}
\alias{substrate}
\title{dataset: substrate}
\format{
data.frame containing data set
  \describe{
  \item{\code{Trat}}{Vector with factor 1}
  \item{\code{bloco}}{Vector with block}
  \item{\code{1,2,3...}}{Numerical vector with germination}
  }
}
\usage{
data("substrate")
}
\description{
The data come from an experiment carried out at the Universidade Estadual de Londrina,
in which four types of substrates were tested in the emergence of sour passion fruit seeds.
The experiment was carried out in a completely randomized design with four replications
of 10 seeds each.
}
\examples{
data(substrate)
}
\keyword{datasets}
