% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tml_function.R
\name{tml}
\alias{tml}
\title{Param: Logistic average time}
\usage{
tml(dados, trat, nrep, time)
}
\arguments{
\item{dados}{Data.frame containing the responses of the evaluations in separate columns side by side and without the columns with the identification of the factors}

\item{trat}{Vector of treatments with n repetitions}

\item{nrep}{Number of repetitions}

\item{time}{Vector containing time}
}
\value{
Returns the vector with the average time.
}
\description{
Param: Logistic average time
}
\examples{
data("substrate")
tml(substrate[,c(3:18)],
      trat = substrate$Trat,
      nrep = 4,
      time = 1:16)
}
