% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loess_function.R
\name{loess_model}
\alias{loess_model}
\title{Analysis: loess regression}
\usage{
loess_model(
  trat,
  resp,
  ylab = "Germination (\%)",
  xlab = expression("Temperature ("^"o" * "C)"),
  theme = theme_classic(),
  error = "SE",
  cardinal = 0,
  width.bar = NA,
  legend.position = "top",
  scale = "none",
  textsize = 12,
  pointsize = 4.5,
  linesize = 0.8,
  pointshape = 21,
  font.family = "sans"
)
}
\arguments{
\item{trat}{Numerical or complex vector with treatments}

\item{resp}{Numerical vector containing the response of the experiment.}

\item{ylab}{Variable response name (Accepts the \emph{expression}() function)}

\item{xlab}{treatments name (Accepts the \emph{expression}() function)}

\item{theme}{ggplot2 theme (\emph{default} is theme_bw())}

\item{error}{Error bar (It can be SE - \emph{default}, SD or FALSE)}

\item{cardinal}{defines the value of y considered extreme (\emph{default} considers 0 germination)}

\item{width.bar}{bar width}

\item{legend.position}{legend position (\emph{default} is c(0.3,0.8))}

\item{scale}{Sets x scale (\emph{default} is none, can be "log")}

\item{textsize}{Font size}

\item{pointsize}{shape size}

\item{linesize}{line size}

\item{pointshape}{format point (\emph{default} is 21)}

\item{font.family}{Font family (\emph{default} is sans)}
}
\value{
\describe{
  \item{\code{Optimum temperature}}{Optimum temperature (equivalent to the maximum point)}
  \item{\code{Optimum temperature response}}{Response at the optimal temperature (equivalent to the maximum point)}
  \item{\code{Minimal temperature}}{Temperature that has the lowest response}
  \item{\code{Minimal temperature response}}{Lowest predicted response}
  \item{\code{Predicted maximum basal value}}{Lower basal limit temperature based on the value set by the user (default is 0)}
  \item{\code{Predicted minimum basal value}}{Upper basal limit temperature based on the value set by the user (default is 0)}
  \item{\code{grafico}}{Graph in ggplot2 with equation}
  }
}
\description{
Fit a polynomial surface determined by one or more numerical predictors, using local fitting.
}
\note{
if the maximum predicted value is equal to the maximum x, the curve does not have a maximum point within the studied range. If the minimum value is less than the lowest point studied, disregard the value.
}
\examples{
library(seedreg)
data("aristolochia")
attach(aristolochia)

#================================
# Germination
#================================
loess_model(trat,germ)

#================================
# Germination speed
#================================
loess_model(trat, vel, ylab=expression("v"~(dias^-1)))
}
\seealso{
\link{loess}
}
\author{
Gabriel Danilo Shimizu

Leandro Simoes Azeredo Goncalves
}
