% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_arrange.R
\name{multicurve}
\alias{multicurve}
\title{Graph: Merge multiple curves into a single graph}
\usage{
multicurve(
  plots,
  theme = theme_classic(),
  legend.title = NULL,
  legend.position = "top",
  trat = NA,
  method = "shape_color",
  fill = "gray90",
  ylab = "Germination (\%)",
  xlab = expression("Temperature ("^"o" * "C)"),
  width.bar = NA,
  pointsize = 4.5,
  linesize = 0.8,
  textsize = 12,
  font.family = "sans"
)
}
\arguments{
\item{plots}{list with objects of type LM_model, BC_model, CD_model, LL_model or normal_model}

\item{theme}{ggplot2 theme (\emph{default} is theme_classi())}

\item{legend.title}{caption title}

\item{legend.position}{legend position (\emph{default} is c(0.3,0.8))}

\item{trat}{name of the curves}

\item{method}{marking method}

\item{fill}{dot fill color in case gray=F}

\item{ylab}{Variable response name (Accepts the \emph{expression}() function)}

\item{xlab}{treatments name (Accepts the \emph{expression}() function)}

\item{width.bar}{bar width}

\item{pointsize}{shape size}

\item{linesize}{line size}

\item{textsize}{Font size}

\item{font.family}{Font family (\emph{default} is sans)}
}
\value{
The function returns a graph joining the outputs of the functions LM_model, LL_model, BC_model, CD_model, loess_model, normal_model, piecewise_model and N_model
}
\description{
Graph: Merge multiple curves into a single graph
}
\details{
The method argument defines the type of markup desired by the user. By default, method="shape_color" is used, which differentiates by color and dot shape. For gray scale, use method="shape_gray". To use only color, use method="color", in this case, the dot shape is 16 (filled circle). You can change the stitch pattern by setting the fill color in quotes followed by a space and the stitch number (eg "gray 21"). Still starting from this last method, if the user uses the change to point format without filling, such as 15, 16, 17 or 18, the function will ignore the first argument (ex. "gray 16"), however, of either way the user must define a color.
}
\examples{
library(seedreg)
data("aristolochia")
attach(aristolochia)
a=LM_model(trat,germ)
b=LL_model(trat,germ,npar = "LL.3")
c=BC_model(trat,germ, npar = "BC.4")
d=CD_model(trat,germ, npar = "CRS.4")
multicurve(list(a,b,c,d))
}
\author{
Gabriel Danilo Shimizu
}
